should be functionally safe and not die
Final addition for the night. For the narrator's style  monogatari.js 2412-2429 should be edited in place of style.css for [data-ui='text'] if ANY character changes a property of the textbox.
Though the permasave/permaload feature should be functional, it still has the issue of loading a step ahead of where it should and I haven't tested it within a clean build, only in my sandbox file where the same issue occured.