Added backup and secret save
The goal of these two features is to enable a load secret save from start option to put a sense of permanence in choices and to prevent data from being lost if mid save a secretsave doesn't copy correctly, as well as provide a backup file in case the player deletes the other saves. This will not protect against cookie/cache deletion the same way that backupdownload does by creating a text file not subject to cache deletion (unless the user clears their downloads folder regularly).